Move the MultiTextField class to a separate file
